Add test background texture to the world.
This is super hacky. The background is a sprite (following this guide:
https://kidscancode.org/blog/2017/04/godot_101_12/) with a manual
location offset in the world and it has been stretched to 16x on both
dimensions and set to repeat as a tiled texture.

The texture itself is complete ass. I generated it with AI, cropped it
in MS Paint, and shipped it.