Add WeaponRing / basis for equipping/aiming weapons.
Just spit-balling a solution here. If the game is top-down and
twin-stick then one option would be for the equipped weapon to "orbit"
the player as they aim. This could work with/without the player visually
rotating as well (or rotating at a different rate, for whatever reason).

WeaponRing doesn't have its own scene and is instead a script attached
to a named node on the Player scene. That one instance currently defines
a debug weapon to visualize the aim direction but it doesn't fire.