Add an empty weapon type, starter weapon to player.
Weapon is totally empty but now exists as a base.

There is a debug_weapon.tscn scene that defines what a weapon could look
like using the appropriate type.

Player has a StarterWeaponTemplate scene that is configurable. Whatever
weapon scene is found there will be equipped to the player's WeaponRing
on start.