Add (basic) knockback support.
This is only for bullet right now. Most of this commit is actually about
debug rendering logic that I added to the Main scene + node extensions.

Player receives a hit (from the bullet when it detects that it has
collided). The hit the player receives includes the original collision
from the bullet.

For the debug rendering. It's all set up by convention.
DebugNodeExtensions.EnableDebugRendering() sets up a /Debug node off of
the main scene, and then populates that node with a debug LineRenderer
control. The knockback code in NonPlayerCharacter.cs calls this via
DebugNodeExtensions.PushLine(...).