Setup basic navmesh support.
World has a NavigationRegion2D child. This defines a polygon that is
(crudely) shaped to fit the world region bounds.

Any colliders that are children of this NavRegion at Bake-time will be
obstacles carved out of the resulting NavMesh.

NonPlayerCharacters have a NavigationAgent2D that
continuously-calculates a path from the current agent position -> the
target position. This is now used to decide where the NPC moves next.