Add a health bar above each shambler.
I tried to do this the idiomatic "Godot" way, which was fun.

NonPlayerCharacter and HealthBar know almost nothing about each other.
NPC has a Godot Signal that emits when the character's health has
changed, and HealthBar has a function on it that matches that Signal's
signature.

The wiring is done in the editor, rather than in code. I connected the
NPC signal to the HealthBar function within shambler.tscn and nowhere
else.

With a bit more work we could easily add this health bar to the player
as well.