Prevent RadialSpawner from running spawn logic inside the editor.
This was my mistake. This type is annotated as a [Tool] so that we can
render debug info in the editor about the size of the inner/outer
radius, but the tool annoation results in _Ready() executing any time
the scene is instantiated (including in the editor) so
dragging/dropping/adding a RadialSpawner anywhere (e.g. test_world.tscn)
caused Shamblers to get spawned *inside the editor*.

This was causing the weird signal cruft to be added on to
test_world.tscn in our commits. It should stop now.

More info on how this works: https://docs.godotengine.org/en/stable/tutorials/plugins/running_code_in_the_editor.html