Make procedural NavMesh size optional, inherit more fields from scene.
Instead of throwing away the entire NavigationPolygon just to resize the
outline, we're now only doing that selectively and keeping the rest of
the polygon that might have been defined in the scene.

So if the scene file says that the Agent Radius computed when the
navmesh bakes should be 25 units, that is what will be used, without us
having to redefine it in code.