Model buildings and their interior regions (rooms).
This was a vertical spike to get a single room (test_room.tscn) built up
with floors, walls, and a roof that can be hidden while the player is
inside.

Buildings have rooms. Each room is defined by an InteriorRegion
(generally a rectangle) which includes an ElevationLevel (int).

Players have a CurrentElevationLevel. Buildings discover the interior
regions they own during start. When a player enters an interior region,
the building determines which regions it owns that should be visibile /
invisible based on the player's CurrentElevationLevel.

Players cannot currently change their elevation level. Technically the
roof of the test building is on level 1 (Ground is level 0), but there's
no way to transition a player between levels yet. That probably looks
like an explicit command to swap their CurrentElevationLevel along with
updating Building visibility, and probably warrants a custom scene node
to facilitate the transfer ("Elevator", "Stairwell", etc.).