Adjust building boundaries to avoid odd collisions.
If the boundary is as big as the building including its walls then the
player can transition inside the building by pushing up against the
outside walls.

Either the boundary needs to be slightly smaller (inside the walls) or
there needs to be a skirt around the whole building. I'm trying the
former here.