Add Interactables.
Place an interactable node on a physics body in the game and it can be
interacted with. There is a single test_interactable.tscn right now and
an instance of it in our test_world.tscn. BodySensor now senses
Interactables in addition to Players and NPCs. If the player walks up to
the test interactable and presses "player_confirm" (Keyboard: E /
Xbox (A) / PS (X)) there is a log message saying who interacted with
what.