Interactable door on test_building.tscn.
This was a lot of work and the initial version here is honestly more
complex than it should be. That said, it was good learning.

The door doesn't have its own scene yet, there's just a door node
implemented in test_building.tscn. The way it works is that the door has
two separate collision bodies:
1. A blocking static body, with a collision layer mask that interacts
   with the player, bullets, etc. This body's collision is entirely
   disabled when the door is "open".
2. A non-blocking static body, placed on collision layer 5 (a new
   "Interactables" layer that the Player now has in its collision mask).
   This is what the player detects with their NearbyBodySensor.

The Door needs both colliders because it can't be disabling the collider
that the player uses to detect the door with in the first place. So they
both exist on separate collision layers instead. It seems to work.