Experiment with making Brains a custom Resource type.
https://docs.godotengine.org/en/stable/tutorials/scripting/resources.html
https://www.reddit.com/r/godot/comments/116naq9/when_to_use_custom_resource_classes_perhaps_with/

^ It seems that there's no harm in defining resources with functional
behavior, so long as we intend for that to be the case. I think Brain
might be a good case for it because it will have its own state that we
want to define common starting values for, and if it updates them at
runtime (and we eventually want to support save games) it's easy to dump
the modified resource to disk. Making Brains resources instead of
something like RefCounted or GodotObject gives us clean ways to define
the Brain parameters inline on NPCs, reference them from common .rs
files, and use inheritance in the editor window (selecting any kind of
Brain to assign).

It also made more sense than making the Brain a child node, in my mind,
because children run _Process() / _PhysicsProcess() *before* their
parents. That makes it hard to have the Brain updating the body. Writing
our own logic for the relationship gives us full control.