Update AI navigation towards enemies to use better nav rules.
This dynamically sets the NavAgent's TargetDesiredDistance based on the
radius of the NPC and its target enemy. It also updates how the AI
interacts with the Brain and NPC a bit. Now:

NPC controls motion (updates to position / rotation based on velocity
and other factors).

Brain controls navigation (updates to velocity and other factors based
on AI instructions).

AI actions instruct the brain on how/where to navigate ("Go to this
enemy", "go to this location").