Add AI Attack action.
Sorta hardcoded at the moment but it works.

AttackEnemyTarget runs continuously once the NPC enters attack range.
The action deactivates itself if attacking is no longer possible.

While the action is active it is marked PausesMotionWhileActive=true,
which means that the Brain will otherwise not attempt to apply new
navigation movement until the attack action is deactivated.