Support sub-360 arcs for RadialSpawner.
Also repositioned this above the buildings and only in a half-circle
now. The problem was that when the buildings overlapped the spawn area
enemies would spawn behind the building and none of their collision
worked correctly.

There might be more to fix there, but this at least gives us more
flexibility in where/how we place enemies in the world.