Add simple pause menu.
Removed debug_quit action and replaced it with a proper pause menu
action that surfaces a scene rendering the menu widgets. While the menu
is open everything else in the scene is paused and all input is handled
by the PauseMenu's _Process() (and button signals).

Options button doesn't work yet but the other two do.