Fix character collision bug
Noticed the player could move through the campfire when changing floors

Shifting the collision mask was overflowing since it includes the world
bound, which is set to the highest bit. This part of the mask wasn't
supposed to be shifted anyways. We probably don't need a special-case world
layer, might revisit that