Introduce Thrower weapon + new Grenade projectile.
Added a new projectile type: Grenade. Grenades can bounce a configurable
number of times and simulate bouncing along the ground due to gravity as
they slow down. (NOTE: This doesn't explode yet).

Added a new Weapon type: Thrower. This is a special weapon instance that
the Player owns one copy of. When a "throwable" is fired, the Thrower is
what does so. Players load the Thrower with a scene template for the
projectile they wish to throw. The primary difference between Thrower
and other weapons is that it allows a charge/wind-up period which alters
the distance the projectiles will travel over time (and for grenades,
this determines where they first "bounce" along the ground).

Thrower gets equipped/unequipped when in use, and the
previously-equipped weapon is re-equipped to the WeaponRing afterward.

This is all sort of stream-of-consciousness so there's tons of room for
improvement. Heavily WIP.