Visualize where throwables will land.
Added a GroundTarget to Thrower that will be positioned near where the
throwable should land/reach based on how long the trigger has been held
down.

Player now exports config for defining the Thrower scene file, which
means we can alter this GroundTarget and other config for the Thrower
independent of the character.

Enemies don't have throwers yet but that's okay.