Make knockback universal to all types of hits.
Moved knockback to be a proper magnitude applied to characters when they
receive a hit as part of the initial HitResult structure. Knockback is
inherent to Projectile children but can be set to 0 to disable the
effect.

Grenade also implements knockback both for the projectile impact itself
and the blast when it explodes.