Refactor impact calculation code.
Made impacts a little simpler to reason about, building off of the last
commit introducing the IImpactMaterial interface.

This changes a fundamental behavior so far: previously, bullets would
collide with characters and do damage to them via
character.ReceiveHit(), which would subsequently decide whether to
render an impact (and which one).

Now, the SOURCE of the impact (e.g. bullet, or explosion) coordinates
the impact - the projectile detects the IImpactMaterial of the recipient
and calculates the impact behavior, including calling
character.ReceiveHit().

The end result is that we only need to teach things which impact other
things how to behave - and can rely on sane defaults for anything
"undefined" that we want to receive an impact. Only characters will
receive damage, but the system can resolve impacts as long as at least
one side of the equation is well-modeled.

Projectiles no longer automatically delete themselves when colliding by
default. It is now controlled by a variable on the projectile
(DestroyOnNextCollision). Grenades set this to false so that they can
bounce along the ground and off of walls before settling/detonating;
Bullets set this to false so that they can bounce off of walls and only
destroy themselves when hitting a soft target; all other projectiles
(e.g. scatter_proj) have this set to true so that they are destroyed on
first impact.