Projectiles can continuously orient toward their velocity.
Cleanup on the Projectile class. Reorganized some of the exported fields
into categories so it's a little cleaner now that there's so many of
them.

Projectiles now can be marked OrientToVelocity=true to automatically
maintain their rotation facing the direction of travel. This is
necessary for pretty much any projectile with a directed shape like an
arrow, missile, etc.

Updated the shape of bullet.tscn to not be a perfect sphere and enabled
this to test/demonstrate.