Fix explosion collision and visibility on different floors
Change the collision layer and mask of the explosion when spawned on
other floors

Make sure the zindex and visibility are updated properly. This isn't
100% ideal atm. Grenades that explode when the player is on another
floors wont be rendered if the player re-enters the floors while they
are mid-explosion. Grenade mid-explosion continue rendering when the
player changes floors. Hoping to refactor/simplify this in the future by
reparenting things inside to the building region instead, which should
also make it easier to fix the aforementioned behavior