Reparent entities inside buildings
Reparent entities that enter building regions to that region. This
simplifies the need to manage updating entity z-index and visiblity for
everything inside a region. Entities are reparented back to the world when
leaving a building

Moved the camera to be a child of main, and configure to follow the
player. This is functionally the same, but avoids an issue where the
camera resets its position when it (or anything above it) gets reparented.
This is also kinda of nice if we want to program events later where the
camera moves focus