Show a player silhouette when hidden behind some objects
This change is experimental, may keep or rollback depending on how it
works out

This change works by moving the player sprite and shader effects into a
viewport - think of this as as separate screen. In this use case, the
viewport is not shown anywhere on screen, so it renders "offscreen".

There are two sprites in the player scene that use this viewport
texture (a flattened image post-shader effects). One is displayed as-is
as zindex 0. The other is displayed at zindex 1, and uses a shader to
compare to original viewport texture - if there are mismatches, it shows
a silhouette color.

This solution requires anything you want to "hide" the player to sit
between the rendering layer of these two sprites, which is a little
inflexible. Since the sprite at zindex 1 is the one ultimately
displayed, anything above this zindex wont create silhouettes when the
player moves underneath it (sometimes desirable, like for the flames of
the campfire, or the clouds in the weather layer). The World scene now contains an empty Node2D for inserting the player, so it's easier to control render order

Lighting only effects the top-level silhouette sprite. Changed daynight
code from using CanvasModulate, otherwise the silhouette would
false-positive on lighting differences. Instead, every entity's shader
handles day/night completely

Also snuck in a small change to model building height and move the player by
a specified offset when jumping off the roof