Fix clouds affecting silhouette when player on roof
https://app.asana.com/1/1209778638119403/project/1209778597616183/task/1209918920181342

The weather zindex was set to 3

The building uses a zindex of 1, and pads the zindex of anything inside
by 1 to provide a "background" layer, e.g. for hiding grenades behind
players. The silhouette sits an additional 1 layer above this - the same
layer as the weather

Set weather layer to 10 so we have some room to add other things at
higher indices