Fix player not being able to shoot when clicking control nodes
Set the mouse filter of visible control nodes to ignore

Part of the problem is we're using a lot of control nodes for
decoration, e.g. in the door and buildings. We should instead use
sprites. Can fix this later or apply it to new structures we add

This is causing problems now because of recent changse to input
handling. With this (recommended) setup, contol nodes will take input
precedence over game input