Make grenade explosions blow up other grenades
Added a config to all projectiles to enable being destroyed by
explosions (just enabled for grenade atm)

This works the same way that grenades detect characters to damage, so
obstructed grenades are not affected

Grenades were originally being set to collision layer 0 when settled so
they didn't collide with anything. Moved them to the interaction layer
so they dont affect physical objects, but could still collide with an
explosion. Debatable if this layer or a new layer should be
used, but this seemed ok for now