Make AI Actions into Resources.
AI Actions are now Resources rather than just RefCounted, they are also
all marked as [GlobalClass] so that we can pick them in the editor and
define them inline on other types (such as the NPC's scene directly
where the Brain is configured).

Renamed the existing attack action to be Melee-specific. Working on a
Ranged variant that's not quite ready yet.

Brain no longer has hardcoded actions. Moved the existing 2 actions to
being defined directly on shambler.tscn.