Allow AI actions to be interrupted.
This is a messy commit but I wanted to get this idea in place without
fleshing out the full secondary attack right away.

AI Actions can now be interrupted if they are written to allow it. This
is controlled by overriding AI.Action.CanInterrupt() to return `true`
whenever the action is interruptible. Right now only MoveToTargetAction
is set to be interruptible so that the AI doesn't mindlessly seek the
target even when they have other actions that they could use from their
current location that are (potentially) scoring higher.

This also checks in some WIP code for a ranged attack action. The action
is configured on Shambler and they will use it whenever the enemy target
is in range and the action is not on cooldown. The ranged attack does
nothing other than spit a log message out about it being used, but the
logs show that the MoveToTarget action is interrupted successfully when
the AI notices that RangedAttack is off cooldown and in-range.