Make Ranged attacks take time, other improvements.
Ranged attacks now take a small amount of time to complete, this is
mostly used to test other interactions at the moment.

Ranged attacks also now require the NPC to be facing its target, though
the checks for this are a bit hastily written at the moment and need
refinement.

Slightly refactored the logic for deciding where the AI is trying to
rotate its body towards now that the AI actions are playing a larger
role there. This is mostly working the way I want, but there's a bit
more to do to support AI tracking targets while they are attacking (as
opposed to being locked at the rotation they started the attack in).

This commit includes some WIP assets and classes (vomit attack,
ParticleSpray.cs) because they are harmless and it makes the NPCs a bit
more interesting to look at while they're cycling actions.

The ranged attack still currently does no damage at all.