Fix bug where AreaEffects would track freed Nearby entities.
This was affecting Explosion but was relevant for all AreaEffects.
Basically the NearbyBodies and NearbyCharacters arrays were updated when
things entered/left, but the elements within these arrays were
references to other bodies that could be freed/deleted at any time. When
this happened (such as an explosion deleting a NearbyCharacter) it was
causing iteration over these arrays to fail casting these invalid
entries to various types.

The solution, for now at least, is double-buffering these arrays.
AreaEffect now throws away the old arrays each tick and replaces them
with new ones populated from the most recent information. This solves
the problem but may be a performance issue down the road. TBD.