Refactor ImpactType names to be less confusing.
Trying to generalize the ImpactTypes as much as possible/reasonable.
Instead of having very bespoke things like "Vomit" and "Human" we now
have general-purpose impact types such as "Sharp", "Bullet", "Blunt" and
so on. This should be a bit easier to reason about when modeling
different impact recipients and how we want them to display damage.

In order to keep this useful and not cumbersome, added a new
CustomUseHint type to the impact type list. This type serves two
purposes:
1. Give things a dedicated symbol to use when they should *not* be
   matched against elements in the recipient's ImpactResponseTable.
2. Causes impact response resolution to completely ignore the
   ImpactResponseTable of the recipient (at least to begin with).

As a concrete example: without this change, the Vomit attack had to pick
an ImpactSourceType from the list. Without coding a bespoke "Vomit"
impact type, the only real option was to pick something imperfect (e.g.
"Burn" or "Bullet") and make sure that the recipient played the correct
impact for that, or otherwise pick "Default" as the ImpactSourceType and
cause the recipient to lookup the default response. Both of these are
bad and will pick the wrong responses almost every time.

Now, Vomit's ImpactSourceType is `CustomUseHint`, and the vomit attack
specifies `impact_vomitsplash` as the DefaultResponseHint to use, and
everything works as expected.