Add decorators for PathMesh.
Added a new concept to PathMesh of a PathDecorator, this is a visitor
pattern that can sample/traverse the PathMesh and generate additional
details along, on top of, or below it.

A big change here is that PathMesh now has more conventions on how its
children are managed:
1. All generated nodes are placed under the "RuntimeGenerated" child
   node.
2. The generated MeshInstance2D for the path is always placed in a
   container called "PathContainer" within the "RuntimeGenerated" child
   node.
3. Nodes which should always draw on top of the path go into a container
   called "AbovePathContainer" within "RuntimeGenerated".
4. Nodes which should always draw beneath the path go into a container
   called "BelowPathContainer" within "RuntimeGenerated".
5. The PathMesh node tracks any nodes added at runtime (via
   AddRuntimeGeneratedChild()) and handles clearing / resetting them
   when the path is regenerated.

A prototype decorator PathDecalDecorator knows how to randomly place
decal sprites along the path mesh. I'm misusing this at the moment to
place barrel sprites, because a barrel was what I found quickly and
looked okay. This decorator would really be better for placing things
the player walks over such as potholes or oil slicks.

Test world is updated to show this off. This is all heavy WIP / POC /
Prototype throwaway work right now.