Support exclusion zones where SmallTown won't place buildings.
Added this so that I could easily exclude the buildings we've
hand-placed and everything else near the conceptual "Start" of the
current test level.

This is a rudimentary check-in. I don't expect this to stay in this
shape beyond the next commit but it's a stepping stone that works and
the next step is to build a custom editor plugin to simplify this a bit.

For the time being, the ScenarioContainer node in our test world has a
RectRegion for the exclusion zone, this is a custom node type that is
then assigned/passed to the SmallTown script so that it knows where it
can't place buildings.