Finished silhouette implementation
Transparent and semi-transparent pixels on the player sprite
are filtered from the silhouette. The latter provides finer detail on
pixel-y art with semi-transparent edges

Added a new animated player sprite to test these changes. This sprite
sheet has animations for 8 directions, so it can be used to setup
directional animation, too