Make RectRegion drag-resizable in editor.
This kinda got crazy but I wanted to see it through once I realized it
worked:

RectRegion is a [Tool] so that it runs in the editor, but most/all of
its functionality ONLY runs in the editor and all it does is expose a
`Rect2` as a draggable, transformable node.

RectRegion is NOT an EditorPlugin. It uses Godot 4+ features to remain a
gameplay component marked as [Tool] while also using the EditorInterface
directly to manipulate selections and input handling in the editor.

At a high level, this works by carefully controlling what nodes are
selected in the editor when input is detected. The secret sauce is
quietly enabling `InternalProcess` notifications for this node, which
DOESN'T cause _Process() to be invoked but DOES cause the notification
to pass through to _Notification() (where this handles input and draw
calls).