Fix many instances of RectRegion not respecting its Transform.
It's hard proxying a type without a Transform (like Rect2) through an
intermediate that does (like RectRegion). I just got a lot of this wrong
before and didn't notice because of the trivial transforms I was testing
with.

This seems far more robust. The exclusion zone in the test world has
been changed to a local position offset and has a non-identity scale
just to prove everything renders correctly.