Make SmallTown place buildings using template.
This uncovered some further issues with the way RectRegions worked when
they had non-identity transforms with position offsets or scales. Most
of that work was made in RectRegion and pushed in the parent commit.

This updates SmallTown to place buildings where they fit, using the
configured BuildingFootprint to test. It also changes the way that
points are filtered out around exclusion regions to test the actual
building footprint rather than an atomic point with 0 size.

Added some rudimentary debug rendering that can be toggled on/off in the
editor on the ScenarioContainer.