Move SmallTown point operations to use IEnumerable.
IEnumerable lets us `yield return VALUE` in our code and defer execution
until the calling code actually needs to iterate the returned list.

I swapped most of the operations in SmallTown to accept and return
IEnumerables because:
1. It seems to be the broad convention for C# code in 2025.
2. It delays execution and avoids unnecessary list conversions as often
   as possible.

Minor performance improvement in exchange for not being able to
(efficiently) count the number of elements in each "list" before
actually converting it to a proper list.