Setup YSorting
Added a YSort node to the main tree that every game object drawn in the
scene goes in. This automatically sorts the tree order based on the Y
position of the sprite, so lower things get drawn above higher things.

This approach simplifies trying to manage the tree order ourselves with
various containers

- Rendering building and campfire on ZIndex 0 instead of 1. Before we would only draw the player behind these, now they can be above or behind depending on position
- Adjusted player collision box to be a small rectangle around the feet
  We might want to implement hitbox and collision separately in the
  future. This lets us see the YSort/depth effect better when standing
  in front of enemies/buildings/campfire, etc
- Enabled YSort on building regions, so YSort works inside buildings