Allow debug draw call groups to be individually enabled/disabled.
Since ProcGen routines run on game startup there's no real opportunity
to use the in-game options to turn the rendering on and off. But there's
no problem (in a dev build at least) with generating all of the debug
data and simply not drawing it unless the option is toggled in-game.

This change enables debug draw call groups to be marked as disabled. The
data they contain is still recorded but the entire drawcall group will
be skipped during rendering. This can be changed at runtime to re-enable
those draw calls.