WIP: Add trees to SmallTown.
This doesn't actually render trees yet but if SmallTown.Trees is enabled
under `Debug Draw Calls` in the options menu there is rendering to
indicate where they would go based on the crude rules in place so far.

Checking this in as-is because I might not get back to it today and it's
a nice chunk to commit.