DebugDrawCallRenderer performance tweaks.
Root caused the issue when enabling Context-aware steering draw calls in
combination with ProcGen routine draw calls: These happen on different
threads, and critically the context-aware steering draw calls happen
within the character's _PhysicsProcess() and NOT during _Process().

The REAL root cause: the single call to `QueueRedraw()` that existed at
the bottom of PushDebugDrawCall(). It's apparently a VERY bad idea to
call QueueRedraw() many times from both the render and physics threads.
You can call it many times, and you can call it many times from EITHER
thread, but calling it many times in BOTH threads is a big
(undocumented) no-no. It simply causes too much thread synchronization
to occur.

This is fine, because it was also unnecessary: We only need to draw once
per frame, so QueueRedraw() is now only called from the renderer's
_Process(). All other calls seem to be fine.

Also introduced a bit of diagnostic information that gets rendered in
tooltips on the "Debug Draw Calls" options menu.