Make WeaponRing aim work with local transforms.
WeaponRing was one of the first bits of this that I wrote and it makes
assumptions about global position and rotation that just aren't always
true.

This change allows the weapon ring to have a local offset (relative to
the player) and still aim in the right direction and position the
equipped weapon in the right location.