Small bugfixes to SmallTown for centered buildings.
We "fixed" this yesterday by adding half the building footprint to each
building's position as we placed it, but I realized that wasn't a proper
fix and only worked by coincidence / because we balanced everything
correctly.

This fixes the core bug, which was that GetPlacementLocations() wasn't
properly considering the PointCloud2D.PointTestOffset value when
calculating the candidate rect to test.

That in turn incovered other minor bugs we'd overlooked yesterday:
the debug rendering for the building placements wasn't using a (now)
centered origin, and the rect data that the building placements produce
(which is fed to Tree placement to avoid dropping trees inside buildings
we just placed) wasn't considering that the placement rects needed
centered origins either.

With all 3 of those fixed, the reusable functions and the SmallTown
routine are both correct now.