Capture more ProcGen data on the placeable scenes.
Previously SmallTown.cs required paremeters like `Footprint` sizes for
both Buildings and Trees. These needed to be consistent with how big the
associated Building and Tree scenes were that got placed, or else things
would not look correct or be placed appropriately.

Things placed by ProcGen now extend Placeable, which is a lightweight
Node2D wrapper that exports details about the Footprint of the placed
scene (and more, eventually).

SmallTown creates a single dummy "prefab" instance of the Building and
Tree scene templates and then interrogates that prefab instance for
details about the footprint when it is placed. The prefab scene is
thrown away / GC'd after SmallTown finishes generation.

Why this approach? Because it lets us have designer tools like
RectRegion available to us in the editor right where we're designing the
trees / buildings / placeable scenes themselves, rather than having to
flip between the placeable scenes and the Scenario scenes (e.g.
SmallTown) to line up values as we change them in one place or another.