Support multiple footprint RectRegions per Placeable.
Placeable now has an enumeration of "SecondaryFootprints" in addition to
the PlacedFootprint. SecondaryFootprints can be checked during ProcGen
routines to determine other criteria. For example, our TestBuilding now
has two SecondaryFootprints: "NoForestRegion" 1 & 2. These regions of
TestBuilding each have the tag "ProcGen.Exclude.Forest", which SmallTown
also uses to look up which placed buildings have secondary footprints
with that tag. This allows SmallTown to not only exclude the building's
PlacedFootprint when determining where to place trees, but also these
additional regions that specify other areas.

Put simply: TestBuilding now has exclusion regions which block the
placement of ProcGen trees in front of the doors and for some extent
beyond that.

Also re-introduced RectRegion.AutoRegisterGroups independent of Tags.
These serve different purposes right now, where Groups are a
globally-queried thing and Tags are local to the instance and can only
be inspected with a reference to a specific instance.

All of this yields a slightly-more-natural looking placement of trees in
SmallTown.