Add basic timing data to SmallTown procgen routines.
Checking in some basic instrumentation because this is starting to slow
down and I want to get ahead of it before it's too slow to use for
anything fun.

Bulk of time spent is on placing trees. This makes sense because there
are many initial points to test (~8k) and each one gets compared against
dozens of rects for each building spread across the map.

There's a limitation in how closely this can time things right now
because the point cloud filters are lazy-evaluated. Basically all of the
time ends up being deferred until the final GetPlacements call that
iterates the whole point cloud to produce the result. More work
necessary here to the filter interface(s) to support getting good
timimgs out.