Massive improvement to SmallTown generation speed.
This is a 10x (!!!) improvement in performance for the tree generation
logic and thus the entire SmallTown generation speed. I am heavily
leveraging IEnumerables and lazy-evaluated sets to limit the number of
times we iterate the same code / how much needs to be cached in memory
to do so. A consequence of that is that sometimes I get it wrong and the
IEnumerable sets we use wind up being iterates many times, and at great
cost (depending on what producing each element in the set requires).

In this case, PlaceableRegionsWithTag() was fetching each node's child
footprint node from the scene by name (slow) as well as fetching another
child node by name (also slow) and then looping over all of its children
to collect them as well (very slow). This slowness was getting
re-executed any time the IEnumerable set produced by this function was
used again. Very Very Very Slow.

Caching the results to lists that can be looped over in ~constant time
is embarassingly efficient by comparison.