Cache Placeable footprints for performance.
Thinking on the last commit a bit more, I don't see a reason for this
state to change during gameplay. We set it up during design in the
editor and then we access it through these getters, so if that's the
full contract there's no reason these can't be cached.

Note regarding the previous commit: I experimented with this change +
reverting the optimization in the last commit. This change to
Placeable.cs on its own brings SmallTown generation times down from ~6
seconds to ~2 seconds. But the optimization in the last commit brings
timings down to ~540ms. So this is more of a general improvement whereas
the last commit was a targeted improvement.