Experiment with more diverse forest in SmallTown.
- Swapped SmallTown to accept a dictionary of tree templates where each
  template is associated with a limit on how many trees of that type can
  be placed (maximally).
- Changed SmallTown forest generation to be multiple passes (1 pass per
  entry in the new dictionary). Passes are ordered by the order of the
  dictionary's entries.
- Added a new test tree with different art, different collision, and
  different footprint. Generating 400 of these in SmallTown currently.

Also disabled the logic that was placing buildings near roads and near
the center of the map. The building placement logic is a toy and not
something I intend to keep around long term. The forest is more
interesting and at the very least we didn't need ~25 buildings on the
map.