Experimenting with tree normal maps.
This isn't really functional, but I used some tools to generate a couple
of normal maps for the main tree sprite. test_tree.tscn uses the normal
map right now but it doesn't do anything in-game unless we put
directional lighting nodes in (or write a shader to work with the
normals, which I would like to do eventually).

This seemed harmless to check in as a starting point though. Also
cleaned up some old test art that we aren't using.