New test building
Added a new test building with some taller height and matching wall
perspectives (front and top visible, 3/4 perspective)

- Building front/facade becomes semi-transparent when the player is
  inside
- Added YSort node to building so it's setup the same way as the outside
  world. The background/floor is not ysorted and sits at ZIndex -1, same
  as the world background, so we dont need to pad the ZIndex of things
  inside a building to have a background layer
- Added some repeating textures to building floors/walls to get an idea of
how it should work. To get textures to line up, wall sizes need to be a
proper multiple of the texture size. I'm using a 64x40 texture which is
kinda a funky size - should probably be something we standardize on in
the future so it's easy/consistent to setup new buildings with different
textures
- The shambler has a little bit of trouble getting inside because the
  door is not as wide. This problem could be solved by switching
  steering to use raycasting instead of the imperfect math I'm currently
  doing. This would also simplify our code. Will do in a future commit
- Didn't setup impact materials for the new walls, will do in future commit - wanna get this checked in